mapToParameterWithExtent: change type of position argument to size_t
This is a more natural type for a position.

+++ tc/core/polyhedral/schedule_transforms-inl.h
@@ -49,12 +49,12 @@ template <typename MappingIdType>
 inline detail::ScheduleTree* mapToParameterWithExtent(
     detail::ScheduleTree* root,
     detail::ScheduleTree* tree,
-    int pos,
+    size_t pos,
     MappingIdType id,
     size_t extent) {
   auto band = tree->elemAs<detail::ScheduleTreeElemBand>();
   CHECK(band) << "expected a band, got " << *tree;
-  CHECK_GE(pos, 0) << "dimension underflow";
+  CHECK_GE(pos, 0u) << "dimension underflow";
   CHECK_LT(pos, band->nMember()) << "dimension overflow";
   CHECK_NE(extent, 0) << "NYI: mapping to 0";
